Don't expose the inner workings of the SASL library
The clients shouldn't touch the internals of the library
and shouldn't know what it looks like.
