+++ src/cram-md5/cram-md5.h
@@ -17,12 +17,9 @@
 #ifndef SRC_CRAM_MD5_CRAM_MD5_H_
 #define SRC_CRAM_MD5_CRAM_MD5_H_ 1
 
-#include "config.h"
-#include "cbsasl/cbsasl.h"
-
 #define MECH_NAME_CRAM_MD5 "CRAM-MD5"
 
-cbsasl_error_t cram_md5_server_init();
+cbsasl_error_t cram_md5_server_init(void);
 
 cbsasl_error_t cram_md5_server_start(cbsasl_conn_t* conn);
